+++ model/article.go
@@ -1,0 +1,5 @@
+package model
+
+type ArticleRequest struct {
+	OrderNo string `json:"orderNo"`
+}
