+++ model/article_list.go
@@ -1,0 +1,8 @@
+package model
+
+type ArticleListRequest struct {
+	ReturnType string `json:"field"`
+	ShowPrice  bool   `json:"showPrice"`
+	ShowStock  bool   `json:"showStock"`
+	MediaType  byte   `json:"mediaType,omitempty"`
+}
