+++ model/article_status.go
@@ -1,0 +1,5 @@
+package model
+
+type ArticleStatusRequest struct {
+	OrderNo string `json:"orderNo"`
+}
