+++ model/authenticate.go
@@ -1,0 +1,5 @@
+package model
+
+type AuthenticateRequest struct {
+	APIKey string `json:"apiKey"`
+}
