+++ model/file_import.go
@@ -1,0 +1,10 @@
+package model
+
+type FileImportRequest struct {
+	File      []byte `json:"file"`
+	FileType  string `json:"fileType,omitempty"`
+	DataType  byte   `json:"dataType,omitempty"`
+	MediaType byte   `json:"mediaType,omitempty"`
+	FormatID  int    `json:"formatId,omitempty"`
+	Encoding  string `json:"encoding,omitempty"`
+}
