+++ model/file_status.go
@@ -1,0 +1,5 @@
+package model
+
+type FileStatusRequest struct {
+	Filename string `json:"filename"`
+}
