+++ model/import_status.go
@@ -1,0 +1,3 @@
+package model
+
+// Does not need any parameters currently
