+++ model/order_cancel.go
@@ -1,0 +1,5 @@
+package model
+
+type OrderCancelRequest struct {
+	OrderID string `json:"orderId"`
+}
