+++ model/order_item_cancel.go
@@ -1,0 +1,6 @@
+package model
+
+type OrderItemCancelRequest struct {
+	OrderItemID string `json:"orderItemId"`
+	MediaType   string `json:"mediaType"`
+}
