+++ model/order_message.go
@@ -1,0 +1,7 @@
+package model
+
+type OrderMessageRequest struct {
+	OrderID        string `json:"orderId"`
+	MessageType    string `json:"messageType"`
+	AdditionalText string `json:"additionalText,omitempty"`
+}
