+++ model/order_status.go
@@ -1,0 +1,6 @@
+package model
+
+type OrderStatusRequest struct {
+	OrderID string `json:"orderId"`
+	Status  string `json:"status"`
+}
